Replace reader objects with functions
Replace object_reader with next_object() and token_reader with next_token().
TokenReader took ownership of a Source trait object.
Since it was necessary to access that trait object from PdfDocument, it was simpler to
simply pass around references to it rather than use Rc<Source> to allow multiple owners.
Since that was the only item stored in a TokenReader and the ObjectReader contained
only a TokenReader, the need for those objects disappeared.